ENH: Replace old "scripted" module template by the new "scripteddesigner" template
We had two scripted module templates: scripted and scripteddesigner. The "scripted" template had several limitations: GUI widgets are created from code, parameter node is not used (user choices are not saved in the scene), logic is not created in the widget (so it is not accessible from other modules).

Removed the old scripted module template to make things simpler (and added an example of a dynamically added node selector widget).
