+++ Utilities/Templates/Modules/Scripted/TemplateKey.py
@@ -1,8 +1,9 @@
 import os
 import unittest
+import logging
 import vtk, qt, ctk, slicer
 from slicer.ScriptedLoadableModule import *
-import logging
+from slicer.util import VTKObservationMixin
 
 #
 # TemplateKey
@@ -15,123 +16,202 @@ class TemplateKey(ScriptedLoadableModule):
 
   def __init__(self, parent):
     ScriptedLoadableModule.__init__(self, parent)
-    self.parent.title = "TemplateKey" # TODO make this more human readable by adding spaces
-    self.parent.categories = ["Examples"]
-    self.parent.dependencies = []
-    self.parent.contributors = ["John Doe (AnyWare Corp.)"] # replace with "Firstname Lastname (Organization)"
+    self.parent.title = "TemplateKey"  # TODO: make this more human readable by adding spaces
+    self.parent.categories = ["Examples"]  # TODO: set categories (folders where the module shows up in the module selector)
+    self.parent.dependencies = []  # TODO: add here list of module names that this module requires
+    self.parent.contributors = ["John Doe (AnyWare Corp.)"]  # TODO: replace with "Firstname Lastname (Organization)"
     self.parent.helpText = """
 This is an example of scripted loadable module bundled in an extension.
-It performs a simple thresholding on the input volume and optionally captures a screenshot.
-"""
-    self.parent.helpText += self.getDefaultModuleDocumentationLink()
+"""  # TODO: update with short description of the module
+    self.parent.helpText += self.getDefaultModuleDocumentationLink()  # TODO: verify that the default URL is correct or change it to the actual documentation
     self.parent.acknowledgementText = """
-This file was originally developed by Jean-Christophe Fillion-Robin, Kitware Inc.
+This file was originally developed by Jean-Christophe Fillion-Robin, Kitware Inc., Andras Lasso, PerkLab,
 and Steve Pieper, Isomics, Inc. and was partially funded by NIH grant 3P41RR013218-12S1.
-""" # replace with organization, grant and thanks.
+"""  # TODO: replace with organization, grant and thanks.
 
 #
 # TemplateKeyWidget
 #
 
-class TemplateKeyWidget(ScriptedLoadableModuleWidget):
+class TemplateKeyWidget(ScriptedLoadableModuleWidget, VTKObservationMixin):
   """Uses ScriptedLoadableModuleWidget base class, available at:
   https://github.com/Slicer/Slicer/blob/master/Base/Python/slicer/ScriptedLoadableModule.py
   """
 
+  def __init__(self, parent=None):
+    """
+    Called when the user opens the module the first time and the widget is initialized.
+    """
+    ScriptedLoadableModuleWidget.__init__(self, parent)
+    VTKObservationMixin.__init__(self)  # needed for parameter node observation
+    self.logic = None
+    self._parameterNode = None
+
   def setup(self):
+    """
+    Called when the user opens the module the first time and the widget is initialized.
+    """
     ScriptedLoadableModuleWidget.setup(self)
 
-    # Instantiate and connect widgets ...
+    # Load widget from .ui file (created by Qt Designer)
+    uiWidget = slicer.util.loadUI(self.resourcePath('UI/TemplateKey.ui'))
+    self.layout.addWidget(uiWidget)
+    self.ui = slicer.util.childWidgetVariables(uiWidget)
 
-    #
-    # Parameters Area
-    #
+    # Set scene in MRML widgets. Make sure that in Qt designer
+    # "mrmlSceneChanged(vtkMRMLScene*)" signal in is connected to each MRML widget's.
+    # "setMRMLScene(vtkMRMLScene*)" slot.
+    uiWidget.setMRMLScene(slicer.mrmlScene)
+
+    # Example of adding widgets dynamically (without Qt designer).
+    # This approach is not recommended, but only shown as an illustrative example.
     parametersCollapsibleButton = ctk.ctkCollapsibleButton()
-    parametersCollapsibleButton.text = "Parameters"
+    parametersCollapsibleButton.text = "More"
+    parametersCollapsibleButton.collapsed = True
     self.layout.addWidget(parametersCollapsibleButton)
-
-    # Layout within the dummy collapsible button
     parametersFormLayout = qt.QFormLayout(parametersCollapsibleButton)
+    self.invertedOutputSelector = slicer.qMRMLNodeComboBox()
+    self.invertedOutputSelector.nodeTypes = ["vtkMRMLScalarVolumeNode"]
+    self.invertedOutputSelector.addEnabled = True
+    self.invertedOutputSelector.removeEnabled = True
+    self.invertedOutputSelector.noneEnabled = True
+    self.invertedOutputSelector.setMRMLScene(slicer.mrmlScene)
+    self.invertedOutputSelector.setToolTip("Result with inverted threshold will be written into this volume")
+    parametersFormLayout.addRow("Inverted output volume: ", self.invertedOutputSelector)
 
-    #
-    # input volume selector
-    #
-    self.inputSelector = slicer.qMRMLNodeComboBox()
-    self.inputSelector.nodeTypes = ["vtkMRMLScalarVolumeNode"]
-    self.inputSelector.selectNodeUponCreation = True
-    self.inputSelector.addEnabled = False
-    self.inputSelector.removeEnabled = False
-    self.inputSelector.noneEnabled = False
-    self.inputSelector.showHidden = False
-    self.inputSelector.showChildNodeTypes = False
-    self.inputSelector.setMRMLScene( slicer.mrmlScene )
-    self.inputSelector.setToolTip( "Pick the input to the algorithm." )
-    parametersFormLayout.addRow("Input Volume: ", self.inputSelector)
+    # Create a new parameterNode
+    # This parameterNode stores all user choices in parameter values, node selections, etc.
+    # so that when the scene is saved and reloaded, these settings are restored.
+    self.logic = TemplateKeyLogic()
+    self.ui.parameterNodeSelector.addAttribute("vtkMRMLScriptedModuleNode", "ModuleName", self.moduleName)
+    self.setParameterNode(self.logic.getParameterNode())
 
-    #
-    # output volume selector
-    #
-    self.outputSelector = slicer.qMRMLNodeComboBox()
-    self.outputSelector.nodeTypes = ["vtkMRMLScalarVolumeNode"]
-    self.outputSelector.selectNodeUponCreation = True
-    self.outputSelector.addEnabled = True
-    self.outputSelector.removeEnabled = True
-    self.outputSelector.noneEnabled = True
-    self.outputSelector.showHidden = False
-    self.outputSelector.showChildNodeTypes = False
-    self.outputSelector.setMRMLScene( slicer.mrmlScene )
-    self.outputSelector.setToolTip( "Pick the output to the algorithm." )
-    parametersFormLayout.addRow("Output Volume: ", self.outputSelector)
+    # Connections
+    self.ui.parameterNodeSelector.connect('currentNodeChanged(vtkMRMLNode*)', self.setParameterNode)
+    self.ui.applyButton.connect('clicked(bool)', self.onApplyButton)
 
-    #
-    # threshold value
-    #
-    self.imageThresholdSliderWidget = ctk.ctkSliderWidget()
-    self.imageThresholdSliderWidget.singleStep = 0.1
-    self.imageThresholdSliderWidget.minimum = -100
-    self.imageThresholdSliderWidget.maximum = 100
-    self.imageThresholdSliderWidget.value = 0.5
-    self.imageThresholdSliderWidget.setToolTip("Set threshold value for computing the output image. Voxels that have intensities lower than this value will set to zero.")
-    parametersFormLayout.addRow("Image threshold", self.imageThresholdSliderWidget)
+    # These connections ensure that whenever user changes some settings on the GUI, that is saved in the MRML scene
+    # (in the selected parameter node).
+    self.ui.inputSelector.connect("currentNodeChanged(vtkMRMLNode*)", self.updateParameterNodeFromGUI)
+    self.ui.outputSelector.connect("currentNodeChanged(vtkMRMLNode*)", self.updateParameterNodeFromGUI)
+    self.ui.imageThresholdSliderWidget.connect("valueChanged(double)", self.updateParameterNodeFromGUI)
+    self.ui.invertOutputCheckBox.connect("toggled(bool)", self.updateParameterNodeFromGUI)
+    self.invertedOutputSelector.connect("currentNodeChanged(vtkMRMLNode*)", self.updateParameterNodeFromGUI)
 
-    #
-    # check box to trigger taking screen shots for later use in tutorials
-    #
-    self.enableScreenshotsFlagCheckBox = qt.QCheckBox()
-    self.enableScreenshotsFlagCheckBox.checked = 0
-    self.enableScreenshotsFlagCheckBox.setToolTip("If checked, take screen shots for tutorials. Use Save Data to write them to disk.")
-    parametersFormLayout.addRow("Enable Screenshots", self.enableScreenshotsFlagCheckBox)
-
-    #
-    # Apply Button
-    #
-    self.applyButton = qt.QPushButton("Apply")
-    self.applyButton.toolTip = "Run the algorithm."
-    self.applyButton.enabled = False
-    parametersFormLayout.addRow(self.applyButton)
-
-    # connections
-    self.applyButton.connect('clicked(bool)', self.onApplyButton)
-    self.inputSelector.connect("currentNodeChanged(vtkMRMLNode*)", self.onSelect)
-    self.outputSelector.connect("currentNodeChanged(vtkMRMLNode*)", self.onSelect)
-
-    # Add vertical spacer
-    self.layout.addStretch(1)
-
-    # Refresh Apply button state
-    self.onSelect()
+    # Initial GUI update
+    self.updateGUIFromParameterNode()
 
   def cleanup(self):
-    pass
+    """
+    Called when the application closes and the module widget is destroyed.
+    """
+    self.removeObservers()
 
-  def onSelect(self):
-    self.applyButton.enabled = self.inputSelector.currentNode() and self.outputSelector.currentNode()
+  def setParameterNode(self, inputParameterNode):
+    """
+    Adds observers to the selected parameter node. Observation is needed because when the
+    parameter node is changed then the GUI must be updated immediately.
+    """
+
+    if inputParameterNode:
+      self.logic.setDefaultParameters(inputParameterNode)
+
+    # Set parameter node in the parameter node selector widget
+    wasBlocked = self.ui.parameterNodeSelector.blockSignals(True)
+    self.ui.parameterNodeSelector.setCurrentNode(inputParameterNode)
+    self.ui.parameterNodeSelector.blockSignals(wasBlocked)
+
+    if inputParameterNode == self._parameterNode:
+      # No change
+      return
+
+    # Unobserve previusly selected parameter node and add an observer to the newly selected.
+    # Changes of parameter node are observed so that whenever parameters are changed by a script or any other module
+    # those are reflected immediately in the GUI.
+    if self._parameterNode is not None:
+      self.removeObserver(self._parameterNode, vtk.vtkCommand.ModifiedEvent, self.updateGUIFromParameterNode)
+    if inputParameterNode is not None:
+      self.addObserver(inputParameterNode, vtk.vtkCommand.ModifiedEvent, self.updateGUIFromParameterNode)
+    self._parameterNode = inputParameterNode
+
+    # Initial GUI update
+    self.updateGUIFromParameterNode()
+
+  def updateGUIFromParameterNode(self, caller=None, event=None):
+    """
+    This method is called whenever parameter node is changed.
+    The module GUI is updated to show the current state of the parameter node.
+    """
+
+    # Disable all sections if no parameter node is selected
+    self.ui.basicCollapsibleButton.enabled = self._parameterNode is not None
+    self.ui.advancedCollapsibleButton.enabled = self._parameterNode is not None
+    if self._parameterNode is None:
+      return
+
+    # Update each widget from parameter node
+    # Need to temporarily block signals to prevent infinite recursion (MRML node update triggers
+    # GUI update, which triggers MRML node update, which triggers GUI update, ...)
+
+    wasBlocked = self.ui.inputSelector.blockSignals(True)
+    self.ui.inputSelector.setCurrentNode(self._parameterNode.GetNodeReference("InputVolume"))
+    self.ui.inputSelector.blockSignals(wasBlocked)
+
+    wasBlocked = self.ui.outputSelector.blockSignals(True)
+    self.ui.outputSelector.setCurrentNode(self._parameterNode.GetNodeReference("OutputVolume"))
+    self.ui.outputSelector.blockSignals(wasBlocked)
+
+    wasBlocked = self.invertedOutputSelector.blockSignals(True)
+    self.invertedOutputSelector.setCurrentNode(self._parameterNode.GetNodeReference("OutputVolumeInverse"))
+    self.invertedOutputSelector.blockSignals(wasBlocked)
+
+    wasBlocked = self.ui.imageThresholdSliderWidget.blockSignals(True)
+    self.ui.imageThresholdSliderWidget.value = float(self._parameterNode.GetParameter("Threshold"))
+    self.ui.imageThresholdSliderWidget.blockSignals(wasBlocked)
+
+    wasBlocked = self.ui.invertOutputCheckBox.blockSignals(True)
+    self.ui.invertOutputCheckBox.checked = (self._parameterNode.GetParameter("Invert") == "true")
+    self.ui.invertOutputCheckBox.blockSignals(wasBlocked)
+
+    # Update buttons states and tooltips
+    if self._parameterNode.GetNodeReference("InputVolume") and self._parameterNode.GetNodeReference("OutputVolume"):
+      self.ui.applyButton.toolTip = "Compute output volume"
+      self.ui.applyButton.enabled = True
+    else:
+      self.ui.applyButton.toolTip = "Select input and output volume nodes"
+      self.ui.applyButton.enabled = False
+
+  def updateParameterNodeFromGUI(self, caller=None, event=None):
+    """
+    This method is called when the user makes any change in the GUI.
+    The changes are saved into the parameter node (so that they are restored when the scene is saved and loaded).
+    """
+
+    if self._parameterNode is None:
+      return
+
+    self._parameterNode.SetNodeReferenceID("InputVolume", self.ui.inputSelector.currentNodeID)
+    self._parameterNode.SetNodeReferenceID("OutputVolume", self.ui.outputSelector.currentNodeID)
+    self._parameterNode.SetParameter("Threshold", str(self.ui.imageThresholdSliderWidget.value))
+    self._parameterNode.SetParameter("Invert", "true" if self.ui.invertOutputCheckBox.checked else "false")
+    self._parameterNode.SetNodeReferenceID("OutputVolumeInverse", self.invertedOutputSelector.currentNodeID)
 
   def onApplyButton(self):
-    logic = TemplateKeyLogic()
-    enableScreenshotsFlag = self.enableScreenshotsFlagCheckBox.checked
-    imageThreshold = self.imageThresholdSliderWidget.value
-    logic.run(self.inputSelector.currentNode(), self.outputSelector.currentNode(), imageThreshold, enableScreenshotsFlag)
+    """
+    Run processing when user clicks "Apply" button.
+    """
+    try:
+      self.logic.run(self.ui.inputSelector.currentNode(), self.ui.outputSelector.currentNode(),
+        self.ui.imageThresholdSliderWidget.value, self.ui.invertOutputCheckBox.checked)
+      if self.invertedOutputSelector.currentNode():
+        # If additional output volume is selected then result with inverted threshold is written there
+        self.logic.run(self.ui.inputSelector.currentNode(), self.invertedOutputSelector.currentNode(),
+          self.ui.imageThresholdSliderWidget.value, not self.ui.invertOutputCheckBox.checked, showResult=False)
+    except Exception as e:
+      slicer.util.errorDisplay("Failed to compute results: "+str(e))
+      import traceback
+      traceback.print_exc()
+
 
 #
 # TemplateKeyLogic
@@ -147,56 +227,45 @@ class TemplateKeyLogic(ScriptedLoadableModuleLogic):
   https://github.com/Slicer/Slicer/blob/master/Base/Python/slicer/ScriptedLoadableModule.py
   """
 
-  def hasImageData(self,volumeNode):
-    """This is an example logic method that
-    returns true if the passed in volume
-    node has valid image data
+  def setDefaultParameters(self, parameterNode):
     """
-    if not volumeNode:
-      logging.debug('hasImageData failed: no volume node')
-      return False
-    if volumeNode.GetImageData() is None:
-      logging.debug('hasImageData failed: no image data in volume node')
-      return False
-    return True
+    Initialize parameter node with default settings.
+    """
+    if not parameterNode.GetParameter("Threshold"):
+      parameterNode.SetParameter("Threshold", "50.0")
+    if not parameterNode.GetParameter("Invert"):
+      parameterNode.SetParameter("Invert", "false")
 
-  def isValidInputOutputData(self, inputVolumeNode, outputVolumeNode):
-    """Validates if the output is not the same as input
+  def run(self, inputVolume, outputVolume, imageThreshold, invert=False, showResult=True):
     """
-    if not inputVolumeNode:
-      logging.debug('isValidInputOutputData failed: no input volume node defined')
-      return False
-    if not outputVolumeNode:
-      logging.debug('isValidInputOutputData failed: no output volume node defined')
-      return False
-    if inputVolumeNode.GetID()==outputVolumeNode.GetID():
-      logging.debug('isValidInputOutputData failed: input and output volume is the same. Create a new volume for output to avoid this error.')
-      return False
-    return True
-
-  def run(self, inputVolume, outputVolume, imageThreshold, enableScreenshots=0):
-    """
-    Run the actual algorithm
+    Run the processing algorithm.
+    Can be used without GUI widget.
+    :param inputVolume: volume to be thresholded
+    :param outputVolume: thresholding result
+    :param imageThreshold: values above/below this threshold will be set to 0
+    :param invert: if True then values above the threshold will be set to 0, otherwise values below are set to 0
+    :param showResult: show output volume in slice viewers
     """
 
-    if not self.isValidInputOutputData(inputVolume, outputVolume):
-      slicer.util.errorDisplay('Input volume is the same as output volume. Choose a different output volume.')
-      return False
+    if not inputVolume or not outputVolume:
+      raise ValueError("Input or output volume is invalid")
 
     logging.info('Processing started')
 
     # Compute the thresholded output volume using the Threshold Scalar Volume CLI module
-    cliParams = {'InputVolume': inputVolume.GetID(), 'OutputVolume': outputVolume.GetID(), 'ThresholdValue' : imageThreshold, 'ThresholdType' : 'Above'}
-    cliNode = slicer.cli.run(slicer.modules.thresholdscalarvolume, None, cliParams, wait_for_completion=True)
-
-    # Capture screenshot
-    if enableScreenshots:
-      self.takeScreenshot('TemplateKeyTest-Start','MyScreenshot',-1)
+    cliParams = {
+      'InputVolume': inputVolume.GetID(),
+      'OutputVolume': outputVolume.GetID(),
+      'ThresholdValue' : imageThreshold,
+      'ThresholdType' : 'Above' if invert else 'Below'
+      }
+    cliNode = slicer.cli.run(slicer.modules.thresholdscalarvolume, None, cliParams, wait_for_completion=True, update_display=showResult)
 
     logging.info('Processing completed')
 
-    return True
-
+#
+# TemplateKeyTest
+#
 
 class TemplateKeyTest(ScriptedLoadableModuleTest):
   """
@@ -229,18 +298,38 @@ class TemplateKeyTest(ScriptedLoadableModuleTest):
     """
 
     self.delayDisplay("Starting the test")
-    #
-    # first, get some data
-    #
+
+    # Get/create input data
+
     import SampleData
-    SampleData.downloadFromURL(
-      nodeNames='FA',
-      fileNames='FA.nrrd',
-      uris='http://slicer.kitware.com/midas3/download?items=5767',
-      checksums='SHA256:12d17fba4f2e1f1a843f0757366f28c3f3e1a8bb38836f0de2a32bb1cd476560')
+    inputVolume = SampleData.downloadFromURL(
+      nodeNames='MRHead',
+      fileNames='MR-Head.nrrd',
+      uris='https://github.com/Slicer/SlicerTestingData/releases/download/MD5/39b01631b7b38232a220007230624c8e',
+      checksums='MD5:39b01631b7b38232a220007230624c8e')[0]
     self.delayDisplay('Finished with download and loading')
 
-    volumeNode = slicer.util.getNode(pattern="FA")
+    inputScalarRange = inputVolume.GetImageData().GetScalarRange()
+    self.assertEqual(inputScalarRange[0], 0)
+    self.assertEqual(inputScalarRange[1], 279)
+
+    outputVolume = slicer.mrmlScene.AddNewNodeByClass("vtkMRMLScalarVolumeNode")
+    threshold = 50
+
+    # Test the module logic
+
     logic = TemplateKeyLogic()
-    self.assertIsNotNone( logic.hasImageData(volumeNode) )
-    self.delayDisplay('Test passed!')
+
+    # Test algorithm with non-inverted threshold
+    logic.run(inputVolume, outputVolume, threshold, True)
+    outputScalarRange = outputVolume.GetImageData().GetScalarRange()
+    self.assertEqual(outputScalarRange[0], inputScalarRange[0])
+    self.assertEqual(outputScalarRange[1], threshold)
+
+    # Test algorithm with inverted threshold
+    logic.run(inputVolume, outputVolume, threshold, False)
+    outputScalarRange = outputVolume.GetImageData().GetScalarRange()
+    self.assertEqual(outputScalarRange[0], inputScalarRange[0])
+    self.assertEqual(outputScalarRange[1], inputScalarRange[1])
+
+    self.delayDisplay('Test passed')
