--- Utilities/Templates/Modules/ScriptedDesigner/Testing/CMakeLists.txt
@@ -1,1 +1,0 @@
-add_subdirectory(Python)
